src/pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "paddlepaddle>=2.6.0",
    "paddleocr>=3.3.2",
    "pdf2image>=1.17.0",
    "pillow>=9.0.0",
    "numpy>=1.24.0",
    "beautifulsoup4>=4.12.0",
]
